pyproject.toml
[project]
name = "elasticGPT"
version = "0.1.0"
description = "Generate questions for the Elastic Search Engineering Exam"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Ajay Bhargava", email = "bhargava.ajay@gmail.com" }
]
dependencies = [
    "baml-py>=0.90.2",
    "elasticsearch>=9.0.2",
    "ipython>=9.3.0",
    "pydantic>=2.11.7",
    "pytest",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "typer>=0.16.0",
]

[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["elasticGPT"]
package-dir = {"elasticGPT" = "src/elasticGPT"}

[tool.ruff]
src = ["src"]
line-length = 100
select = ["E", "F", "B", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["__test__"]
addopts = "-ra -q"
